Remove duplication by using fixtures in HTTP server tests.
